Renamed chat/index.* to chat/chat.*
