+++ src/components/chat/chat.tsx
@@ -5,8 +5,8 @@ import DocumentModel from '../../model/document/document-model';
 import LocalSettings from '../../model/local/local-settings';
 import { Header } from '../general-elements/header/header';
 import ScrollContainer from '../scroll-container/scroll-container';
+import './chat.scss';
 import ChatMessageEditor from './editor/editor';
-import './index.scss';
 import { MessageBubble } from './message-bubble/message-bubble';
 
 export function ChatPage({
