+++ src/components/chat/message-bubble/message-bubble.cy.tsx
@@ -6,7 +6,7 @@ import '../../../main.scss';
 import Author from '../../../model/document/author';
 import DocumentModel from '../../../model/document/document-model';
 import Message from '../../../model/document/message';
-import '../index.scss';
+import '../chat.scss';
 import MessageBubble from './message-bubble';
 
 const anyWindow: any = window;
